requirements.txt
beautifulsoup4==4.6.3
boto==2.49.0
boto3==1.9.71
botocore==1.12.71
bs4==0.0.1
bz2file==0.98
certifi==2018.11.29
chardet==3.0.4
dateparser==0.7.0
docutils==0.14
gensim==3.6.0
idna==2.8
jieba==0.39
jmespath==0.9.3
lxml==4.2.5
mongo==0.2.0
numpy==1.15.4
pandas==0.23.4
psutil==5.4.8
psycopg2==2.7.6.1
pymongo==3.7.2
python-dateutil==2.7.5
python-Levenshtein==0.12.0
pytz==2018.7
redis==3.0.1
regex==2018.11.22
requests==2.21.0
s3transfer==0.1.13
scipy==1.2.0
selenium==3.8.0
six==1.12.0
smart-open==1.7.1
threadpool==1.3.2
tzlocal==1.5.1
urllib3==1.24.1
